refactor: share postgres predefined migrations between db-postgres and db-vercel-postgres
