+++ packages/db-postgres/src/exports/migration-utils.ts
@@ -1,1 +1,1 @@
-export { migratePostgresV2toV3 } from '../predefinedMigrations/v2-v3/index.js'
+export { migratePostgresV2toV3 } from '@payloadcms/drizzle/postgres'
