--- packages/db-vercel-postgres/src/predefinedMigrations/v2-v3/types.ts
@@ -1,9 +1,0 @@
-/**
- * Set of all paths which should be moved
- * This will be built up into one WHERE query
- */
-export type PathsToQuery = Set<string>
-
-export type DocsToResave = {
-  [id: number | string]: Record<string, unknown>[]
-}
