+++ packages/drizzle/src/exports/postgres.ts
@@ -10,5 +10,6 @@ export { execute } from '../postgres/execute.js'
 export { getMigrationTemplate } from '../postgres/getMigrationTemplate.js'
 export { init } from '../postgres/init.js'
 export { insert } from '../postgres/insert.js'
+export { migratePostgresV2toV3 } from '../postgres/predefinedMigrations/v2-v3/index.js'
 export { requireDrizzleKit } from '../postgres/requireDrizzleKit.js'
 export * from '../postgres/types.js'
